Import Rcpp
Previously valr had Rcpp in LinkingTo, but not in Imports

This meant that the Rcpp name space was not loaded when valr was loaded,
and consequently causes an error in valr.

    Error in bed12toexons_impl(x) :
       function 'enterRNGScope' not provided by package 'Rcpp'
     Calls: bed12_to_exons -> bed12toexons_impl

This issue was masked previous to readr 1.4.0 because readr had Rcpp in
its Imports, but the Rcpp dependency was removed in readr 1.4.0.