Updates for readr 2.0.0
in readr 2.0.0 newlines at the end of a file will generate a row of NA
data, rather than being silently ignored. So we needed to modify the
test data to avoid a test failure due to this.

Additionally because readr 2.0.0 uses lazy parsing by default the
warning would not be triggered during the initial read, but instead
later when the malformed value is read. Setting `lazy = FALSE` instead
reads the file eagerly and preserves the expected behavior.